chore: move ApplyRuntimeDefaults to pkg/merge — out of api/v1alpha1
